+++ pom.xml
@@ -57,7 +57,23 @@
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>io.springfox</groupId>
+      		<artifactId>springfox-boot-starter</artifactId>
+      		<version>3.0.0-SNAPSHOT</version>
+		</dependency>
 	</dependencies>
+	
+	<repositories>
+    	<repository>
+      		<id>jfrog-snapshots</id>
+		  	<name>JFROG Snapshots</name>
+      		<url>http://oss.jfrog.org/artifactory/oss-snapshot-local</url>
+      		<snapshots>
+        		<enabled>true</enabled>
+      		</snapshots>
+    	</repository>
+  	</repositories>
 
 	<build>
 		<plugins>
