+++ src/main/java/com/shorty/app/utilis/NameOfServer.java
@@ -1,4 +1,4 @@
-package com.shorty.project.help;
+package com.shorty.app.utilis;
 
 public class NameOfServer {
 	//Change this variable to your public IP in order to host on the Internet
