+++ src/main/java/com/shorty/app/utilis/RandomString.java
@@ -1,4 +1,4 @@
-package com.shorty.project.help;
+package com.shorty.app.utilis;
 
 import java.util.Random;
